django_nextjs_template/backend/pyproject.toml
[tool.poetry]
name = "<project_name>"
version = "0.2.0"
description = "<project_name>"
package-mode = false
authors = ["greyone"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
Django = "5.1.6"
gunicorn = "23.*"
uvicorn = "0.34.*"
sentry-sdk = "2.22.*"
django-stubs-ext = "5.1.*"
django-storages = "^1.14.4"
django-ninja = "1.3.0"
boto3 = "^1.35.23"
psycopg = {extras = ["binary"], version = "^3.2.2"}

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[tool.isort]
profile = "black"

[tool.black]
skip-string-normalization = true
exclude = ".*/migrations/.*"

[tool.pytest.ini_options]
addopts = "--reuse-db --no-migrations"
DJANGO_SETTINGS_MODULE = "tests.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
